requirements.txt
python-telegram-bot>=21.0
google-api-python-client
google-auth
apscheduler
python-dotenv
pytz
google-generativeai
Pillow
flask
